requirements.txt
beautifulsoup4==4.6.0
certifi==2018.1.18
minio==3.0.4
Pillow==5.0.0
python-dateutil==2.7.0
pytz==2018.3
six==1.11.0
Unipath==1.1
urllib3==1.22
